feat: detach reusable functionality from QuickLaneOverview
