+++ src/shared/constants/quick-lane.ts
@@ -1,0 +1,10 @@
+export const QUICK_LANE_COLUMNS = {
+	TITLE: 'title',
+	URL: 'url',
+	CREATED_AT: 'created_at',
+	UPDATED_AT: 'updated_at',
+	CONTENT_LABEL: 'content_label',
+	AUTHOR: 'author',
+};
+
+export const ITEMS_PER_PAGE = 20;
